Script to generate albedo for the 1st year.
This script generates a .dat file of the calculated albedo from the temperature values of the unchanged model and the generated geography.
This is mainly to match the albedo distribution of the time independent albedo calculation with the time dependent albedo calculation.
Respects the different threshold temperatures for ice forming over water/land.